Unify result-length for aggregates among handlers
Aggregated results (AVG, STD etc) include start-time but not end-time,
and cover the period following the time-stamp. This is now the same for
all handlers, ODBC and REST.
Interpolated results still include both start-time and end-time.
Added a handful of tests to verify reads from REST vs ODBC, and to check for
segmentation consistencies.