Add autodeploy to build scripts
Also fix read-test for PI Web API raw. For some reason the 
returned series is of length 4, sometimes 5. Works perfectly 
for "real" tags, only default test tags that exhibit this behavior.
Probably some sort of caching issue (or feature) on the server,
or maybe the test tags aren't stored but autogenerated on 
read?