Fix bug where tags w/o data were not returned
Some tags with no data in the time range specified would return a
completely empty dataframe.
This has been changed so that a column should always be returned.
However, sometimes it will return columns consisting of None, sometimes
columns of NaN and sometimes empty columns. This depends on the handler,
whether the tag is queried alongside other tags that do have data etc.
At least for now it is up to the user to handle these various cases
appropriately.