Update web_handlers.py (#181)
* Update web_handlers.py

Hi, was trying to find how to handle LIKE statements in the SQL query via the REST API and came across your great library. I notice that you don't current have some code to parse SQL queries into a dataframe so have suggested some code...

* Update web_handlers.py

Added processing of SQL query to repalce tabs and new lines with spaces that are typical of formatted SQL queries.  Note that LIKE statments do not seem to be compatible with the requests. I suspect this is to do with the % symbol not being passed. See my Stack Overflow question https://stackoverflow.com/questions/76063389/percent-symbol-in-python-requests-causing-query-to-not-work?noredirect=1#comment134147387_76063389

---------

Co-authored-by: Åsmund Våge Fannemel <34712686+asmfstatoil@users.noreply.github.com>